requirements.txt
insightface
onnxruntime
opencv-python
tqdm
numpy
scikit-learn
matplotlib
torch
huggingface-hub
transformers
safetensors
pillow
facenet-pytorch
tensorflow
h5py
torchvision
dlib-bin
omegaconf
einops
timm

# ---Library Program 01---
# insightface
# onnxruntime
# opencv-python
# tqdm

# ---Library Program 02---
# numpy
# opencv-python
# tqdm
# torch
# huggingface-hub
# transformers
# safetensors
# pillow
# dlib
# facenet-pytorch
# tensorflow
# h5py
# torchvision
# onnxruntime
# omegaconf
# einops
# timm

# ---Library Program 03---
# numpy
# scikit-learn
# matplotlib